Connect log viewer to LogsView. Todo: service log streaming.
