+++ lib/javascript/mdc/src/MDCDrawerReact.js
@@ -12,6 +12,10 @@ export class MDCDrawerReact extends React.Component {
     this.list = new MDCList(this.refManager.refs.mainList);
     this.list.wrapFocus = true;
 
+    if (this.props.items.length >= 1) {
+      this.list.selectedIndex = 0;
+    }
+
     this.list.listen("MDCList:action", (e) => {
       if (this.props.action) {
         this.list.selectedIndex = e.detail.index;
@@ -37,12 +41,14 @@ export class MDCDrawerReact extends React.Component {
         elements.push(
           <a key={x} className="mdc-list-item">
             <span className="mdc-list-item__ripple" />
-            <i
-              className="material-icons mdc-list-item__graphic"
-              aria-hidden="true"
-            >
-              {this.props.items[x].icon}
-            </i>
+            {this.props.items[x].icon && (
+              <i
+                className="material-icons mdc-list-item__graphic"
+                aria-hidden="true"
+              >
+                {this.props.items[x].icon}
+              </i>
+            )}
             <span className="mdc-list-item__text">
               {this.props.items[x].label}
             </span>
